requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
requests
google-generativeai